refactor: replace "dataAPI" file from "pages". "dataAPI.ts" shouldn`t be in this folder, because it is not a page. During a building process "pages" directory is checked if each file have export-component. If its haven`t, then build will fail, thus I have moved this file and change its` path in imports.
