+++ .idea/workspace.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="ChangeListManager">
-    <list default="true" id="fec91a7a-2410-477f-acd1-05e38d1ed1ab" name="Changes" comment="fix: code formatting.&#10;in setting file: &lt;label&gt; should be a parent for the checkbox to be connected with it. Other way it will trigger linter.">
+    <list default="true" id="fec91a7a-2410-477f-acd1-05e38d1ed1ab" name="Changes" comment="fix: code formatting">
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/pages/signin.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/pages/signin.tsx" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/pages/signup.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/pages/signup.tsx" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/pages/dataAPI.ts" beforeDir="false" afterPath="$PROJECT_DIR$/API/dataAPI.ts" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/pages/index.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/pages/index.tsx" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -57,6 +57,8 @@
       <recent name="$PROJECT_DIR$/public" />
     </key>
     <key name="MoveFile.RECENT_KEYS">
+      <recent name="C:\Users\Lanna\Desktop\innopoisk\API" />
+      <recent name="C:\Users\Lanna\Desktop\innopoisk" />
       <recent name="$PROJECT_DIR$/public" />
       <recent name="$PROJECT_DIR$/pages" />
     </key>
@@ -140,7 +142,14 @@
       <option name="project" value="LOCAL" />
       <updated>1657289957444</updated>
     </task>
-    <option name="localTasksCounter" value="8" />
+    <task id="LOCAL-00008" summary="fix: code formatting">
+      <created>1657290179741</created>
+      <option name="number" value="00008" />
+      <option name="presentableId" value="LOCAL-00008" />
+      <option name="project" value="LOCAL" />
+      <updated>1657290179741</updated>
+    </task>
+    <option name="localTasksCounter" value="9" />
     <servers />
   </component>
   <component name="TypeScriptGeneratedFilesManager">
@@ -165,6 +174,7 @@
     <MESSAGE value="fix: build errors connected to husky" />
     <MESSAGE value="fix: code formatting, ESlint rules" />
     <MESSAGE value="fix: code formatting.&#10;in setting file: &lt;label&gt; should be a parent for the checkbox to be connected with it. Other way it will trigger linter." />
-    <option name="LAST_COMMIT_MESSAGE" value="fix: code formatting.&#10;in setting file: &lt;label&gt; should be a parent for the checkbox to be connected with it. Other way it will trigger linter." />
+    <MESSAGE value="fix: code formatting" />
+    <option name="LAST_COMMIT_MESSAGE" value="fix: code formatting" />
   </component>
 </project>
